Polished docstrings, ran goimports and swapped fmt.Errorf with errors.New where possible
